:sparkles: add tick management to the server
